Remove unused AlreadyDonatedModal
- the banners use a simple link/"button" to trigger a close/maybelater event currently

- rename the AlreadyDonatedModal feature tests to AlreadyDonatedLink and remove all tests around
the former modal

https://phabricator.wikimedia.org/T384044
