+++ banners/english/C25_WMDE_Desktop_EN_00/messages.ts
@@ -4,7 +4,6 @@ import UpgradeToYearlyEn from '@src/components/DonationForm/Forms/messages/Upgra
 import SoftCloseEn from '@src/components/SoftClose/messages/SoftClose.en';
 import FooterEn from '@src/components/Footer/messages/Footer.en';
 import MainDonationFormEn from '@src/components/DonationForm/Forms/messages/MainDonationForm.en';
-import AlreadyDonatedModalEn from '@src/components/AlreadyDonatedModal/translations/AlreadyDonatedModal.en';
 import FallbackBanner from '@src/components/FallbackBanner/messages/FallbackBanner.en';
 import DoubleProgressBarEn from '@src/components/ProgressBar/messages/DoubleProgressBar.en';
 
@@ -14,9 +13,9 @@ const messages: TranslationMessages = {
 	...SoftCloseEn,
 	...FooterEn,
 	...MainDonationFormEn,
-	...AlreadyDonatedModalEn,
 	...FallbackBanner,
 	...DoubleProgressBarEn,
+	'already-donated-link': 'I\'ve already donated',
 	'upgrade-to-yearly-copy': '<p>Every year we are dependent on the support of people like you. Yearly donations ' +
 		'help sustainably and enable long term development.</p>' +
 		'<p>No risks attached, you can tell us to stop at any time.</p>'
