+++ banners/pad/C25_WMDE_iPad_DE_00/styles/styles.scss
@@ -49,5 +49,4 @@
 @use '@src/themes/Treedip/Footer/SelectionInput';
 @use '@src/themes/Treedip/Message/Message';
 @use '@src/themes/Treedip/Slider/KeenSlider';
-@use '@src/themes/Treedip/AlreadyDonatedModal/AlreadyDonatedModal';
 @use '@src/themes/Treedip/SoftClose/SoftClose';
