--- src/components/AlreadyDonatedModal/AlreadyDonatedModal.scss
@@ -1,11 +1,0 @@
-@mixin layout {
-	&-already-donated {
-		position: absolute;
-		height: 100%;
-		width: 100%;
-		display: flex;
-		align-items: center;
-		justify-content: center;
-		top: 0;
-	}
-}
