--- src/components/AlreadyDonatedModal/translations/AlreadyDonatedModal.de.ts
@@ -1,7 +1,0 @@
-const Translations = {
-	'already-donated-open-link': 'Habe schon gespendet',
-	'already-donated-maybe-later-button': 'Vielleicht später',
-	'already-donated-go-away-button': 'Für dieses Jahr reicht es'
-};
-
-export default Translations;
