--- src/components/AlreadyDonatedModal/translations/AlreadyDonatedModal.en.ts
@@ -1,7 +1,0 @@
-const Translations = {
-	'already-donated-open-link': 'I\'ve already donated',
-	'already-donated-maybe-later-button': 'Maybe later',
-	'already-donated-go-away-button': 'Enough for this year'
-};
-
-export default Translations;
