--- src/themes/Svingle/AlreadyDonatedModal/AlreadyDonatedModal.scss
@@ -1,63 +1,0 @@
-@use '../variables/globals';
-@use '../variables/fonts';
-@use 'src/components/AlreadyDonatedModal/AlreadyDonatedModal';
-
-.wmde-banner {
-	@include AlreadyDonatedModal.layout;
-
-	&-already-donated {
-		visibility: hidden;
-		opacity: 0;
-		transition: opacity 300ms globals.$banner-easing, visibility 300ms globals.$banner-easing;
-		z-index: 9999;
-		background: rgba( 0, 0, 0, 0.6 );
-		font-size: 18px;
-		text-align: center;
-		font-family: fonts.$content;
-
-		&-content {
-			transform: scale( 1.1 );
-			position: relative;
-			max-width: 800px;
-			margin: 0 10px;
-			background: var( --main-background );
-			border-radius: 6px;
-			padding: 30px 30px 20px;
-			transition: transform 300ms globals.$banner-easing;
-		}
-
-		&-buttons {
-			text-align: center;
-			margin-top: 30px;
-		}
-
-		&-button {
-			display: inline-block;
-			border: 0 none;
-			min-width: 220px;
-			color: var( --already-donated-button-color );
-			background-color: var( --already-donated-button-background );
-			border-radius: 18px;
-			font-size: 14px;
-			line-height: 32px;
-			font-weight: bold;
-			text-align: center;
-			margin: 0 16px 10px;
-			cursor: pointer;
-			transition: border-radius 300ms globals.$banner-easing;
-
-			&:hover,
-			&:focus {
-				border-radius: 8px;
-			}
-		}
-
-		&--is-visible {
-			visibility: visible;
-			opacity: 1;
-			.wmde-banner-already-donated-content {
-				transform: scale( 1 );
-			}
-		}
-	}
-}
