+++ test/banners/english/C25_WMDE_Desktop_EN_00/components/BannerVar.spec.ts
@@ -14,7 +14,7 @@ import { useFormModel } from '@src/components/composables/useFormModel';
 import { resetFormModel } from '@test/resetFormModel';
 import { bannerAutoHideFeatures, bannerMainFeatures } from '@test/features/MainBanner';
 import { DynamicContent } from '@src/utils/DynamicContent/DynamicContent';
-import { alreadyDonatedModalFeatures } from '@test/features/AlreadyDonatedModal';
+import { alreadyDonatedLinkFeatures } from '@test/features/AlreadyDonatedLink';
 import { softCloseFeatures } from '@test/features/SoftCloseDesktop';
 import { Timer } from '@src/utils/Timer';
 import { TimerStub } from '@test/fixtures/TimerStub';
@@ -180,7 +180,7 @@ describe( 'BannerVar.vue', () => {
 		test.each( [
 			[ 'expectFiresMaybeLaterEventOnLinkClick' ]
 		] )( '%s', async ( testName: string ) => {
-			await alreadyDonatedModalFeatures[ testName ]( getWrapper() );
+			await alreadyDonatedLinkFeatures[ testName ]( getWrapper() );
 		} );
 	} );
 } );
